% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomSkew.R
\name{BinomSkew}
\alias{BinomSkew}
\title{Skewness in the Binomial Family of Distributions}
\usage{
BinomSkew()
}
\value{
no value.  Graphical side-effects only.
}
\description{
An app to investigate how skewness in a binomial distribtution vanishes
when np is large enough.  Sample size is set at n = 50, but the user can vary p with a slider.
}
\examples{
\dontrun{
if (require(manipulate)) BinomSkew()
}
}
\author{
Homer White (hwhite0@georgetowncollege.edu)
}
