% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabrera}
\alias{cabrera}
\title{Miguel Cabrera}
\format{
A data frame with 6265 observations on the following 12 variables.  Each observation
is a pitch to Cabrera.
\describe{ \item{season}{The year of play}
\item{gamedate}{Date of the game in which the pitch was thrown}
\item{pitch_type}{Type of pitch thrown, as determined by a computer algorithm.}
\item{balls}{Current ball count}
\item{strikes}{Current strike count}
\item{speed}{speed of pitch (in mph).  (When crossing plate?)}
\item{px}{x-coordinate of pitch (in feet, measured from center of plate)}
\item{pz}{vertical coordinate of pitch (in feet above plate)}
\item{swung}{Whether or not Cabrera swung at the ball.  Factor with levels "no", "yes".}
\item{hitx}{x-coordinate of landing point of ball (if it was hit).  Relative to park.}
\item{hity}{y-coordinate of landing point of ball (if it was hit).  Relative to park.}
\item{hit_outcome}{Outcome when ball was hit.  Factor with levels E (error), H (hit), O (batter out).}
}
}
\source{
Marchi and Albert:  analyzing Baseball Data with R, CRC Press 2014.  For
more on the PITCHf/x system, see \url{https://en.wikipedia.org/wiki/PITCHf/x}.
}
\description{
PITCHf/x data on nine-time All-Star Miguel Cabrera, who is thought to be one of the best
pure hitters in baseball.  Covers seasons 2009 through 2012.
}
\keyword{datasets}
