% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chugtime}
\alias{chugtime}
\title{Time to Chug}
\format{A data frame with 13 observations on the following 2 variables.
\describe{
\item{Weight}{Weight of the subject (in pounds).}
\item{ChugTime}{How long (in seconds) the subject requires to drink the
beverage.}
}}
\source{
Utts and Heckard, Mind on Statistics, 4th Edition.
}
\description{
College-aged males chugging a 12-ounce can of a certain beverage.
}
\keyword{datasets}

