% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIProp.R
\name{CIProp}
\alias{CIProp}
\title{Confidence Intervals (for one population proportion)}
\usage{
CIProp()
}
\value{
Graphical and numerical output
}
\description{
An app to investigate how many times a confidence interval for one population proportion
 captures the true population parameter. The true population proportion is plotted as a vertical red line and
 the user can visually see how changes to the sample, population proportion, sample size, and confidence level
 affect the width of the confidence interval.  Summary information is output to the console to tally the number of times
 the computed confidence interval covers the true population mean and how many times
 it misses.
}
\note{
Uses manipulate from R Studio
}
\examples{
\dontrun{
if (require(manipulate)) CIProp()
}
}
\author{
Rebekah Robinson \email{rebekah_robinson@georgetowncollege.edu}
}
