% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DtrellScat}
\alias{DtrellScat}
\title{Dynamic Trellising (Scatterplot)}
\usage{
DtrellScat(form,data)
}
\arguments{
\item{form}{A formula of the form \code{y~x|c}.  All three variables in the formula
should be from the data frame \code{data}.  \code{c} May be a factor or numerical.}

\item{data}{A data frame.}
}
\value{
Graphical and numerical output.
}
\description{
An app to facilitate exploration of the relationship between two numerical variables, conditonal upon
the values of a third variable.
}
\examples{
\dontrun{
if (require(manipulate)) DtrellScat(sat~salary|frac,data=sat)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

