% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expCounts.R
\name{expCounts}
\alias{expCounts}
\title{Expected Cell Counts From Two-Way Tables}
\usage{
expCounts(tab)
}
\arguments{
\item{tab}{A table with two dimensions, or an object that can be coerced to one.}
}
\description{
Simple instructional function to compute expected cell counts 
from a table of observed counts.
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

