% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeSchool.R
\docType{data}
\name{FakeSchool}
\alias{FakeSchool}
\title{An Imaginary School}
\format{
A data frame with 28 observations on the following 5 variables.
\describe{ \item{Students}{Name of each student}
\item{Sex}{sex of the student}
\item{class}{class rank of the student}
\item{GPA}{grade poitn average}
\item{Honors}{whether or not he student is in the Honors Program} }
}
\description{
Hypothetical school, used for illustrative purposes
}
\keyword{datasets}
