% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel}
\alias{fuel}
\title{Speed and Fuel Efficiency (British Ford Escort)}
\format{A data frame with 15 observations on the following 2 variables.
\describe{
\item{speed}{in kilometers per hour.}
\item{efficiency}{fuel efficiency, measured in liters of fuel required to travel 100 kilometers. }
}}
\source{
The Basic Practice of Statistics, by Moore and McCabe.
}
\description{
A British Ford Escort was driven along a prescribed course.  Each drive was
done at a different speed, and the fuel efficiency was recorded for each
drive.
}
\keyword{datasets}

