% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galton}
\alias{galton}
\title{Galton's Father-Son Data}
\format{
A data frame with 1078 observations on the following 2 variables.
\describe{
\item{fheight}{Height of the father, in inches.}
\item{sheight}{Height of the son, in inches.}
}
}
\description{
Data on father-son pairs.  Collected in 1885 by Francis Galton.
}
\keyword{datasets}
