% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{gcstudents}
\alias{gcstudents}
\title{Georgetown College Students}
\format{A data frame with 62 observations on the following 4 variables.
\describe{ \item{height}{height of the survey participant, in inches}
\item{GPA}{grade-point average}
\item{enough_Sleep}{Does the particpant feel that he/she gets enough sleep?}
\item{sex}{sex of the survey participant}
}}
\source{
MAT 111 at Georgetown College
}
\description{
Data collected by GC students.
}
\keyword{datasets}

