% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss02}
\alias{gss02}
\title{General Social Survey, 2002}
\format{
A data frame with 2765 observations on the following 13 variables.
\describe{
\item{sex}{a factor with levels \code{Female}
\code{Male}}
\item{race}{a factor with levels \code{AfrAm}
\code{Hispanic} \code{Other} \code{White}}
\item{degree}{a factor
with levels \code{Bachelor} \code{Graduate} \code{HighSchool} \code{JunColl}
\code{NotHs}}
\item{relig}{a factor with levels \code{Catholic}
\code{Jewish} \code{Other} \code{Protestant}}
\item{polparty}{a
factor with levels \code{Democrat} \code{Independent} \code{Other}
\code{Republican}}
\item{cappun}{a factor with levels \code{Favor}
\code{Oppose} Whether or not the subject favors capital punishment.}
\item{tvhours}{the subject estimates number of
hours per day he or she watches TV.}
\item{marijuan}{a factor with
levels \code{Legal} \code{NotLegal} Whether or not subject believes that
marijuana should be legalized.}
\item{owngun}{a factor with levels
\code{No} \code{Yes}.  Does the subject own a gun?}
\item{gunlaw}{a factor with levels
\code{Favor} \code{Oppose} Whether or not the subject favors stricter
gunlaws.}
\item{age}{age of the subject}
\item{chldidel}{the ideal number of children the subject would like to have.}
\item{emailtime}{estimated number of hours per week subject spends using email.}
}
}
\source{
National Opinion Research Center:  \url{http://www3.norc.org/gss+website/}.  
Found in Uts and Heckard: Mind on Statistics, 4th Edition.
}
\description{
The General Social Survey (GSS) is a nationwide poll that has been conducted
since 1972 (semiannually since 1994).  Most interviews are done
face-to-face.  For further information, see below.
}
\keyword{datasets}
