% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss08}
\alias{gss08}
\title{General Social Survey, 2008}
\format{
A data frame with 2023 observations on the following 12 variables.
\describe{ \item{sex}{a factor with levels \code{Female}
\code{Male}} \item{race}{a factor with levels \code{AfrAm}
\code{Other} \code{White}} \item{degree}{a factor with levels
\code{Bachelor} \code{Graduate} \code{HighSchool} \code{JunColl}
\code{NotHs}} \item{relig}{a factor with levels \code{Catholic}
\code{Jewish} \code{None} \code{Other} \code{Protestant}}
\item{polparty}{a factor with levels \code{Democrat}
\code{Independent} \code{Other} \code{Republican}} \item{cappun}{a
factor with levels \code{Favor} \code{Oppose}} \item{tvhours}{a
numeric vector} \item{marijuan}{a factor with levels \code{Legal}
\code{NotLegal}} \item{owngun}{a factor with levels \code{No}
\code{Yes}} \item{gunlaw}{a factor with levels \code{Favor}
\code{Oppose}} \item{age}{a numeric vector}
\item{chldidel}{a numeric vector} }
}
\source{
National Opinion Research Center:  \url{https://gss.norc.org/}.  
Found in Uts and Heckard: Mind on Statistics, Fourth Edition.
}
\description{
General Social Survey, 2008
}
\references{
For more information see \code{gss02}
}
\keyword{datasets}
