\docType{data}
\name{hair_and_act}
\alias{hair_and_act}
\title{Hair Color and ACT Score}
\format{A data frame with 100 observations on the following 3 variables.
\describe{
\item{sex}{a factor with levels \code{female} \code{male}}
\item{hair.color}{a factor with levels \code{dark} \code{light}}
\item{act}{composite ACT score of
subject.}
}}
\source{
  MAT 111 at Georgetown College
}
\description{
  A study performed by MAT 111 students at Georgetown
  College.
}
\keyword{datasets}

