% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{handheight}
\alias{handheight}
\title{Height and Handspan}
\format{A data frame with 167 observations on the following 3 variables.
\describe{
\item{Sex}{a factor with levels \code{Female} \code{Male}}
\item{Height}{height of subject, in inches.}
\item{HandSpan}{handspan of subject, in centimeters.}
}}
\source{
Uts and Heckard, Mind on Statistics, 4th Edition.
}
\description{
Height and handspan of a few subjects.
}
\keyword{datasets}

