% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanford1}
\alias{hanford1}
\title{Handford Weather Station, 1984-2010}
\format{A data frame with 27 observations on the following 2 variables.
\describe{
\item{year}{calendar year}
\item{temp}{average high temperature for that year.}
}}
\source{
For more on the Hanford station, see \url{http://www.hanford.gov/page.cfm/HMS}
}
\description{
The station is located in Hanford, WA.
}
\keyword{datasets}

