% Generated by roxygen2 (4.0.1): do not edit by hand
\name{helpGC}
\alias{helpGC}
\title{Quick Vignettes in the Viewer}
\usage{
helpGC(topic,package="tigerstats")
}
\arguments{
\item{topic}{filename of the vignette, exclusive of the .html extension}

\item{package}{Name of the installed package containing the desried vignette}
}
\value{
side effects
}
\description{
A convenience function to show package vignettes in the R Studio Viewer pane.
}
\examples{
\dontrun{
helpGC(lmGC)
}
}
\author{
Homer White (hwhite0@georgetowncollege.edu)
}

