% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hofbatting}
\alias{hofbatting}
\title{Hall of Fame Batters}
\format{
A data frame with 147 observations on the following 29 variables.
\describe{ \item{Name}{Player name}
\item{Rk}{Unknown}
\item{Inducted}{Year inducted into Hall of Fame}
\item{Yrs}{Number of years planed in the Majors}
\item{From}{First year in the Majors}
\item{To}{Last year in the Majors}
\item{MidCareer}{Middle year of player's career}
\item{Era}{Era of Baseball history in which the player was (for the most part) active.
Values are:  19th century  (before 1900), Dead Ball (1900-1919), Lively Ball (1920-1940),
Integration, (1941-1959), Expansion (1960-1975), Free Agency (1976-1992), Long Ball (1993 +).}
\item{ASG}{Number of All-Star games played}
\item{WAR.pos}{Wins above replacement (WAR), as defined for a position player}
\item{G}{Games played}
\item{PA}{Number of plate appearances}
\item{AB}{Number of times at bat}
\item{R}{Runs scored}
\item{H}{Base hits}
\item{X2B}{Doubles}
\item{X3B}{Triples}
\item{Triple.Rate}{Number of triples divided by number of times at bat}
\item{HR}{Home runs}
\item{HR.Rate}{Number of home runs divided by number of times at bat}
\item{RBI}{Runs batted in}
\item{SB}{Number of successful stolen base attempts}
\item{CS}{Number of times thrown out while attempting to steal a base}
\item{BB}{Base on Balls (number of times "walked")}
\item{SO}{Number of times struck out}
\item{BA}{Batting average}
\item{OBP}{On base percentage}
\item{SLG}{Slugging average}
\item{OPS}{OBP plus SLG}
 }
}
\source{
Modified from Marchi and Albert:  analyzing Baseball Data with R, CRC Press 2014.
}
\description{
Data on the 147 batters inducted into the Major LeagueBaseball Hall of Fame
as of the year 2013.
}
\keyword{datasets}
