\docType{data}
\name{imagpop}
\alias{imagpop}
\title{An Imaginary Population}
\format{A data frame with 10000 rows and 8 variables}
\description{
  An imaginary population, used for instructional purposes.
  The variables are as follows:

  \itemize{ \item sex. (male, female).  \item math. Whether
  or not you were a mathematics major.  \item income.  Your
  annual income, rounded to the nearest $100.  \item
  cappun.  Opinion about the death penalty (favor, oppose).
  \item height.  Height in inches.  \item idealheight.  The
  height you would like to be, in inches.  \item diff.
  Idealheight - actual height.  \item kkardashtemp.  Your
  feelings about Kim Kardashian on a 0-100 scale (0=very
  cold, 100=very warm). }
}
\keyword{datasets}

