\docType{data}
\name{iqsiblings}
\alias{iqsiblings}
\title{IQs of Siblings}
\format{A data frame with 80 observations on the following 2 variables.
\describe{
\item{First}{IQ of the older sibling.}
\item{Second}{IQ of the younger sibling.}
}}
\source{
  William Harris, Georgetown College
}
\description{
  IQs of pairs of siblings.
}
\keyword{datasets}

