% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels}
\alias{labels}
\title{Labels and Perception of Quality}
\format{
A data frame with 30 observations on the following 3 variables.
\describe{
\item{jiffrating}{rating subject gave to the PB in the jar with the Jiff label}
\item{greatvaluerating}{rating subject gave to the PB in the jar with the Great Value label}
\item{sex}{a factor with levels \code{female} \code{male}}
}
}
\source{
MAT 111 at Georgetown College
}
\description{
Students in MAT 111 performed an experiment to see whether the perception of
the quality of peanut butter was affected by the labeling on the peanut
butter jar.  Each subject tasted from two jars, one of which was labeled
Jiff, and the other of which was labeled Great Value (a cheaper brand).
Unknown to the subjects, both jars contained Great Value peanut butter.
Each subject rated the quality of the peanut butter on a scale of 1 to 10.
}
\keyword{datasets}
