% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ledgejump}
\alias{ledgejump}
\title{Crowd Behavior at Ledge-Jumping Incidents}
\format{
A data frame with 21 rows and 2 variables
}
\source{
"The baiting crowd in episodes of threatened suicide",
\emph{Journal of Personality and Social Psychology}, 41, 703-709.  See also dataset 59 in
\emph{A Handbook of Small Datasets} by Hand et al.  See also \doi{10.1037/0022-3514.41.4.703}.
}
\description{
A dataset recreated from summary data describing the relationship between weather and crowd behavior during 21 recorded  
incidents in England, in which a (suicidal) person was contemplating jumping from a ledge or other high structure and a crowd 
gathered to watch.  The variables are as follows:

\itemize{
  \item weather. Warm or cool, based on the time of year when the incident occurred.  
  \item crowd.behavior. The crowd either baited the would-be jumper, or was polite.
}
}
\keyword{datasets}
