% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m111survey}
\alias{m111survey}
\title{MAT 111 Survey}
\format{A data frame with 71 rows and 12 variables}
\source{
Georgetown College, MAT 111.
}
\description{
Results of a survey of MAT 111 students at Georgetown College.

\itemize{
      \item \strong{height}  How tall are you, in inches?
      \item \strong{ideal_ht}  A numeric vector  How tall would you 
      LIKE to be, in inches?
      \item \strong{sleep}  How much sleep did you get last night?
      \item \strong{fastest}  What is the highest speed at which 
      you have ever driven a car?
      \item \strong{weight_feel}  How do you feel about your weight?
      \item \strong{love_first}  Do you believe in love at first sight?
      \item \strong{extra_life}  Do you believe in extraterrestrial life?
      \item \strong{seat}  When you have a choice, where do you prefer 
      to sit in a classroom?
      \item \strong{GPA}  What is your college GPA?
      \item \strong{enough_Sleep}  Do you think you get enough sleep?
      \item \strong{sex}  What sex are you?
      \item \strong{diff.ideal.act.} Your ideal height minus your actual height.
}
}
\keyword{datasets}
