% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m111surveyfa13}
\alias{m111surveyfa13}
\title{MAT 111 Survey, Fall 2013}
\format{A data frame with 85 observations on the following 14 variables.
\describe{
\item{height}{Your height in inches.}
\item{ideal_ht}{How tall you would LIKE to be, in
inches.}
\item{sleep}{How much sleep you got last
night, in hours.}
\item{fastest}{What is the
highest speed at which you have ever driven a car (in mph)?}
\item{weight_feel}{a factor with levels \code{1_underweight}
\code{2_about_right} \code{3_overweight} How do you feel about your weight?}
\item{love_first}{a factor with levels \code{no} \code{yes} Do you
believe in love at first sight?}
\item{extra_life}{a factor with levels
\code{no} \code{yes} Do you believe in life on other planets?}
\item{seat}{a factor with levels \code{1_front}
\code{2_middle} \code{3_back} When you have a choice, where do you prefer to
sit in a classroom?}
\item{GPA}{What is your
current GPA?}
\item{enough_Sleep}{a factor with levels \code{no}
\code{yes} Do you think you get enough sleep?}
\item{sex}{a factor
with levels \code{female} \code{male} What sex are you?}
\item{diff}{ideal height minus actual height}
\item{symbol}{a factor with levels \code{a}
\code{b} (Anchor for the next question.)  For the next question,
either Australia or the US, along with its population, was given in the
leadup information to the question.  The "anchor" variable records which version of the question you were
given.  If "a", the population of Australia was given.  If "b", the U.S. population was given.}
\item{pop_Canada}{"The population of country
XXX is YYY million.  About what is the population of Canada, in millions?"  XXX was either the U.S. or Australia.}
}}
\source{
MAT 111 at Georgetown College
}
\description{
Results of a survey given at beginning of semester, to all students in MAT 111.
}
\keyword{datasets}

