% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{music}
\alias{music}
\title{Music and Reading Comprehension}
\format{
A data frame with 44 observations on 4 variables.
}
\source{
Matt Doolin, MAT 111 at Georgetown College
}
\description{
An experiment performed by a student at Georgetown College.  Forty-four subjects were randomized into four
groups.  All subjects read an article; one group read in a silent environment, while the other three groups
heard each three different genres of music.  Each subject took a reading comprehension test afterward.

\describe{
\item{sex}{a factor with levels \code{Female} \code{Male}}
\item{year}{class rank of subject}
\item{type}{type of music subject listened to while reading}
\item{score}{number of questions correct on reading comprehension test}
}
}
\keyword{datasets}
