\docType{data}
\name{napkins}
\alias{napkins}
\title{Napkin Use}
\format{A data frame with 86 observations on the following 2 variables.
\describe{ \item{napkins}{number of napkins used
by the subject during the meal.} \item{sex}{a factor with levels
\code{female} \code{male} Sex of the person being observed}}}
\source{
MAT 111 at Georgetown College
}
\description{
Students at GC observed their fellow students in the Cafe
at lunch.
}
\examples{
data(napkins)
}
\keyword{datasets}

