% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nosmokeday}
\alias{nosmokeday}
\title{Nicotine Withdrawal and Accidents}
\format{A data frame with 10 observations on the following 3 variables.
\describe{ \item{year}{calendar year}
\item{Injuries.before.NSD}{number of injury accidents on the day one week prior to 
National No Smoke Day in the United Kingdom} 
\item{Injuries.on.NSD}{number of injury accidents on  
National No Smoke Day in the United Kingdom}}}
\source{
J. Knwles, "Nicotine withdrawal and road accidents", Science, 400, 128, (8 July 1999).
Found in Whitlock and Schluter, The Analysis of Biological Data.
}
\description{
Results of study conducted in Great Britain to see if nicotine withdrawal increases the risk of an accident.
}
\keyword{datasets}

