% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{oldfaithful}
\alias{oldfaithful}
\title{Old Faithful}
\format{A data frame with 299 observations on the following 2 variables.
\describe{
\item{Duration}{duration of eruption, in minutes}
\item{TimeNext}{time until the next eruption, in minutes}
}}
\source{
Unknown
}
\description{
Old faithful geyser at Yellowstone Park.
}
\keyword{datasets}

