\docType{data}
\name{ostrichtemp}
\alias{ostrichtemp}
\title{Body and Brasin Temperatures of Ostriches}
\format{A data frame with 6 observations on the following 2 variables.
\describe{ \item{body.temp}{carotid arterial blood temperature of the ostrich (degrees Centigrade).}
\item{brain.temp}{brain temperature (near hypothalumus) of the ostrich (degrees Centigrade).}}}
\source{
  Fuller et. al., "Variability in brain and arterial blood
  temperatures in free-ranging ostriches in their natural
  habitat", The Journal of Experimental Biology 206: pp.
  1171 to 1181.  April 2003.
  \url{http://jeb.biologists.org/content/206/7/1171.long}.
  Found in Whitlock and Schluter, The Analysis of
  Biological Data.
}
\description{
  Body and brain temperatures of six free-ranging
  ostriches.
}
\keyword{datasets}

