% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pennstate1}
\alias{pennstate1}
\title{Penn State #1}
\format{
A data frame with 190 observations on the following 9 variables.
\describe{
\item{Sex}{a factor with levels \code{F} \code{M}}
\item{HrsSleep}{how many hours of sleep the subject gets per night}
\item{SQpick}{a factor
with levels \code{Q} \code{S}.  Each subject was presented with two letters
(S and Q), and asked to pick one.  This variable indicates which letter the
subject picked.}
\item{Height}{height in inches}
\item{RandNumb}{a numeric vector: Each subject was asked to choose
randomly an integer from 1 to 10.}
\item{Fastest}{highest speed, in mph, at which subject has ever driven a car}
\item{RtSpan}{span of the right hand, in
centimeters.}
\item{LftSpan}{span of the left
hand, in centimeters.}
\item{Form}{a factor with levels \code{QorS}
\code{SorQ}.  The order of presentation of the S and Q options to the
subject varied from one survey form to another.  This variable indicates
which letter was presented first on the form.}
}
}
\source{
Uts and Heckard, Mind on Statistics, 4th Edition.
}
\description{
A study of students at Penn State University.
}
\keyword{datasets}
