% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmGC.R
\name{plot.GClm}
\alias{plot.GClm}
\title{Diagnostic Plots for GC Linear Regression}
\usage{
\S3method{plot}{GClm}(x,...)
}
\arguments{
\item{x}{An object of class GClm}

\item{\ldots}{ignored}
}
\value{
two diagmostic plots
}
\description{
Used by generic plot function
}
\examples{
SpeedModel <- lmGC(fastest~GPA,data=m111survey)
plot(SpeedModel)
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

