% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfitGC.R
\name{plot.polyGC}
\alias{plot.polyGC}
\title{Diagnostic Plots for GC Polynomial Regression}
\usage{
\S3method{plot}{polyGC}(x,...)
}
\arguments{
\item{x}{An object of class polyGC}

\item{\ldots}{ignored}
}
\value{
two diagnostic plots
}
\description{
Used by generic plot function
}
\examples{
mpgModel <- polyfitGC(mpg~wt,data=mtcars)
plot(mpgModel)
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
