\name{predict.GClm}
\alias{predict.GClm}
\title{Prediction Function for GC Linear Regression}
\usage{
  \S3method{predict}{GClm}(object,x,...)
}
\arguments{
  \item{object}{An object of class GClm}

  \item{x}{value of the predictor variable}

  \item{\ldots}{ignored}
}
\value{
  numeric prediction
}
\description{
  Used by generic predict function
}
\author{
  Homer White \email{hwhite0@georgetowncollege.edu}
}

