% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polyfitGC.R
\name{predict.polyGC}
\alias{predict.polyGC}
\title{Prediction Function for GC Polynomial Regression}
\usage{
\S3method{predict}{polyGC}(object,x,level=NULL,...)
}
\arguments{
\item{object}{An object of class polyGC}

\item{x}{value of the predictor variable}

\item{level}{desired level of prediction interval}

\item{\ldots}{ignored}
}
\value{
numeric prediction
}
\description{
Used by generic predict function
}
\examples{
#predict mpg for a car weighing 3 tons:
mpgModel <- polyfitGC(mpg~wt,data=mtcars,degree=2)
predict(mpgModel,x=3.0)
#include prediction interval:
predict(mpgModel,x=3.0,level=0.95)
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

