% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.GCbinomtest}
\alias{print.GCbinomtest}
\title{Print Function for binomestGC}
\usage{
\S3method{print}{GCbinomtest}(x,...)
}
\arguments{
\item{x}{An object of class GCbinomtest.}

\item{\ldots}{ignored}
}
\value{
Output to the console.
}
\description{
Utility print function
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

