% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.GCchisqtest}
\alias{print.GCchisqtest}
\title{Print Function for chisqtestGC}
\usage{
\S3method{print}{GCchisqtest}(x,...)
}
\arguments{
\item{x}{An object of class GCchisqtest.}

\item{\ldots}{ignored}
}
\value{
Output to the console and to the plot window
}
\description{
Utility print function
}
\details{
##########################################################
#########################################################
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

