% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmGC.R
\name{print.GClm}
\alias{print.GClm}
\title{Print Function for GC Linear Regression}
\usage{
\S3method{print}{GClm}(x,...)
}
\arguments{
\item{x}{an object of class GClm}

\item{\ldots}{ignored}
}
\value{
graphical output and output to console
}
\description{
Utility print function
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

