\name{print.gcp2test}
\alias{print.gcp2test}
\title{Print Function for GC Proportions Test (Two-Sample)}
\usage{
  \S3method{print}{gcp2test}(x,...)
}
\arguments{
  \item{x}{An object of class gcp2test.}

  \item{\ldots}{ignored}
}
\value{
  Output to the console.
}
\description{
  Utility print function
}
\author{
  Homer White \email{hwhite0@georgetowncollege.edu}
}

