% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestGC.R
\name{print.GCttest}
\alias{print.GCttest}
\title{Print Function for ttestGC}
\usage{
\S3method{print}{GCttest}(x,...)
}
\arguments{
\item{x}{An object of class GCttest.}

\item{\ldots}{ignored}
}
\value{
Output to the console.
}
\description{
Utility print function
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

