% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfitGC.R
\name{print.polyGC}
\alias{print.polyGC}
\title{Print Function for GC Polynomial Regression}
\usage{
\S3method{print}{polyGC}(x,...)
}
\arguments{
\item{x}{an object of class polyGC}

\item{\ldots}{ignored}
}
\value{
graphical output and output to console
}
\description{
Utility print function
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

