% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pushups}
\alias{pushups}
\title{Push-ups by Football Players at Georgetown College}
\format{
A data frame with 30 observations on the following 3 variables.
\describe{
\item{weight}{weight of subject in
pounds.}
\item{pushups}{number of push-ups
completed.}
\item{position}{a factor with levels \code{LINE}
\code{SKILL}: type of position played by the subject.  Line positions
require high body mass, skill positions require a lot of running around.}
}
}
\source{
MAT 111, Georgetown College
}
\description{
Two football players at GC asked their team-mates to do as many push-ups as
they could in two minutes.
}
\keyword{datasets}
