% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_normal_factory.R
\name{random_normal_factory}
\alias{random_normal_factory}
\title{Make Normal Sampling Functions}
\usage{
random_normal_factory(mean, sd)
}
\arguments{
\item{mean}{mean of normal distribution from which to sample}

\item{sd}{standard deviation of the normal distribution}
}
\value{
a function of a single parameter n, with default value 1.
}
\description{
Makes a function that samples from a normal distribution.
}
\examples{
\dontrun{
sampler <- random_normal_factory(mean = 70, sd = 5)
## sample one
sampler()
## sample another
sampler()
## sample a third time
sampler()
## sample another 1000
sampler(n = 1000)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
