\name{RegEstimate}
\alias{RegEstimate}
\title{Estimation of Regression Coefficients}
\usage{
RegEstimate(x=1:10)
}
\arguments{
  \item{x}{A numerical vector, specifying the fixed set of
  x-values.}
}
\value{
Graphical and numerical output.
}
\description{
An app to explore estimation of coefficients in simple
regression.
}
\examples{
\dontrun{
if (require(manipulate)) RegEstimate()
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

