% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saltmarsh}
\alias{saltmarsh}
\title{Effect of Soil Salinity on Plant Growth}
\format{A data frame with 24 observations on the following 3 variables.
\describe{ \item{salt}{amount of salt applied to the plot (in parts per million)}
\item{biomass}{total biomass of plot at the end of the study period (units unknown)}
\item{block}{field in which the plot was located} }}
\source{
The Course Notes of Carl Schwarz, Simon Fraser University:  
\url{http://people.stat.sfu.ca/~cschwarz/CourseNotes/}
}
\description{
Result of an experiment conducted to investigate the effect of salinity level in soil 
on the growth of plants.
}
\details{
From the source (see below):  "Experimental
fields of land were located at an agricultural field station, and each field was divided into six smaller
plots. Each of the smaller plots was treated with a different amount of salt (measured in ppm) and the
biomass at the end of the experiment was recorded."
}
\keyword{datasets}
