\docType{data}
\name{sat}
\alias{sat}
\title{SAT Scores}
\format{A data frame with 50 rows and 8 variables}
\source{
  Deborah Lynn Guber, "Getting what you pay for: the debate
  over equity in public school expenditures" (1999),
  \emph{Journal of Statistics Education} 7(2).
  \url{http://www.amstat.org/publications/jse/secure/v7n2/datasets.guber.cfm}.
}
\description{
  SAT scores by state.  The variables are as follows:

  \itemize{ \item state.  A state in the U.S.  \item
  expend.  Mean annual expenditure per student (in 1000$).
  \item ratio.  Mean student-teacher ratio.  \item salary.
  Mean annual teacher salary.  \item frac.  Percentage of
  students in the state who take the SAT.  \item verbal.
  Mean SAT Verbal score for the state.  \item math.  Mean
  SAT Math score for the state.  \item sat. Sum of mean
  Verbal and mean Math. }
}
\keyword{datasets}

