% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sealsO2}
\alias{sealsO2}
\title{Weddell Seal Oxygen Consumptions}
\format{
A data frame with 10 observations on the following 2 variables.
\describe{ \item{o2.nonfeeding}{Oxygen consumption during recovery time after a dive during which
 no plankton was consumed by the seal, in ml of O2 per kilogram of weight} 
\item{o2.feeding}{Oxygen consumption during recovery time after a dive during which
 plankton was consumed, in ml of O2 per kilogram of weight}}
}
\source{
Williams, T. M., L. A. Fuiman, M. Horning, and R. W. Davis. 2004. 
The Journal of Experimental Biology 207: 973 to 982.  \url{http://jeb.biologists.org/content/207/6/973.full}
}
\description{
Results of an experiment conducted on ten Weddell seals.
}
\keyword{datasets}
