% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{simpleFind}
\alias{simpleFind}
\title{Get a variable from its name}
\usage{
simpleFind(varName,data)
}
\arguments{
\item{varName}{Character string giving the name of the variable to be searched for.}

\item{data}{Usually a data frame that supplies the some or all of the variables for a formula
that is has been passed to the calling function.}
}
\description{
Primitive utility function, for writing functions that handle formula input.  simpleFind
looks first in the environment associated with the data argument.  If nothing is found, it looks
in the parent environment, and so on up the chain.  The intent is to allow use of formula constructed
from names of variables that may not appear in the data frame of interest, but which are present in the
caller's environment (usually the Global Environment).  Functions that use formulas now are more flexible
in an interactive context.

To do:  (1) find a way to make gentler error messages.  (2) as with earlier versions of mosaic, we get
the scoping problem when supplied data frame is named 'labels'.
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

