% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleKeyRev}
\alias{simpleKeyRev}
\title{Modifed Simple Key}
\usage{
simpleKeyRev(text, points = TRUE,
rectangles = FALSE,
lines = FALSE,
col = add.text$col,
cex = add.text$cex,
alpha = add.text$alpha,
font = add.text$font,
fontface = add.text$fontface,
fontfamily = add.text$fontfamily,
lineheight = add.text$lineheight,
...)
}
\arguments{
\item{text}{}

\item{points}{}

\item{rectangles}{}

\item{lines}{}

\item{col}{}

\item{cex}{}

\item{alpha}{}

\item{font}{}

\item{fontface}{}

\item{fontfamily}{}

\item{lineheight}{}

\item{...}{}
}
\description{
Utility function for \code{barchartGC}
}
\keyword{internal}

