% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlowGoodness.R
\name{SlowGoodness}
\alias{SlowGoodness}
\title{Chi Square Resampler (One at a Time) for Goodness-of-Fit}
\usage{
SlowGoodness(x,p)
}
\arguments{
\item{x}{a one-dimensional table, or a vector of observed counts}

\item{p}{vector of null probabilities}
}
\value{
Graphical and numerical output
}
\description{
An app to illustrate use of the chi-square statistic to test
for goodness of fit.  The P-value is computed
by resampling, and the resamples are done one at a time.  A histogram of resampled chi-square statistics is displayed
after each resample, and summary information is output to the
console.
}
\examples{
\dontrun{
throws <- c(one=8,two=18,three=11,four=7,five=9,six=7)
SlowGoodness(throws,p=rep(1/6,6))
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
