% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stumps}
\alias{stumps}
\title{Larvae on Stumps}
\format{A data frame with 23 observations on the following 2 variables.
\describe{
\item{stumps}{number of stumps in the plot}
\item{larvae}{number of larvae clusters in the plot}
}}
\source{
Basic Practice of Statistics, by Moore and McCabe.
}
\description{
Biologists were interested in whether beetles prefer areas where beavers have cut down cottonwood trees.
(The tree-stumps produce tender green shoots that beetles are thought to like.)  23 circular plots, all of equal area,
were studied.  For each plot the researchers counted the number of cottonwood stumps, and also the number of
clusters of beetle larvae found in the plot.
}
\keyword{datasets}
