% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Temperature in U.S. Cities}
\format{
A data frame with 20 observations on 5 variables.
}
\source{
Mind on Statistics, Fourth Edition, Uts and Heckard.
}
\description{
Average temperatures for cities in the United States.

\describe{
\item{city}{Name of the city}
\item{latitude}{latitude of the city, in degrees north of the Equator}
\item{JanTemp}{mean temperature of the city in January.}
\item{AprTemp}{mean temperature of the city in April.}
\item{AugTemp}{mean temperature of the city in August.}
}
}
\keyword{datasets}
