% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themerpres.R
\name{themerpres}
\alias{themerpres}
\title{Lattice Theme or R Presentations}
\usage{
themerpres()
}
\value{
Returns a list to be supplied as the \code{theme} to the \code{lattice} function
\code{\link[lattice]{trellis.par.set}()}.
}
\description{
Modifies the current theme for use with lattice graphics in R Presentation dicuments. Increases size of title,
 axis labels and axis numbers, thickens some lines, etc.
}
\examples{
trellis.par.set(theme=themerpres())

}
\seealso{
\code{\link[lattice]{trellis.par.set}}, \code{\link[lattice]{show.settings}}
}
\keyword{graphics}
