% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tornado}
\alias{tornado}
\title{Tornado Damage}
\format{
A data frame with 51 observations on 2 variables.
}
\source{
Moore and McCabe, The Basic Practice of Statistics.
}
\description{
Tornado damage in the U.S., by state.  Also includes Puerto Rico.

\itemize{ 
\item{\code{state}} {the state}
\item{\code{damage}} {mean annual damage from tornadoes, over a five-year period, in millions of dollars}
}
}
\keyword{datasets}
