% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{udavis2}
\alias{udavis2}
\title{UC Davis #2}
\format{A data frame with 239 observations on 14 variables.}
\source{
Mind on Statistics, 4th edition, Uts and Heckard.
}
\description{
Results of a survey of students at UC-Davis.

\itemize{
\item{\code{Sex}} Sex of the subject
\item{\code{GPA}} Grade point average of the subject
\item{\code{Seat}} Where do you prefer to sit in a classroom, when you have a choice?
\item{\code{alcohol}} How many drinks do you have per week, on average?
\item{\code{WtFeel}} How do you feet about your weight?
\item{\code{Height}} Your height in inches
\item{\code{IdealHt}} Your ideal height
\item{\code{momheight}} Height of your mother
\item{\code{dadheight}} Height of your father
\item{\code{Hand}} What is your dominant hand?
\item{\code{Looks}} When it comes to judging a person as a potential mate, how important are looks?
\item{\code{Friends}} With whom do you make friends more easily:  people of the same sex or of the opposite sex?
\item{\code{Cheat}} Do you cheat on exams?
\item{\code{Smoke}} Are you a smoker?
}
}
\keyword{datasets}

