% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{verlander}
\alias{verlander}
\title{Justin Verlander}
\format{A data frame with 15307 observations on the following 12 variables.  Each observation
is a single pitch.
\describe{ \item{season}{The year of play}
\item{gamedate}{Date of the game in which the pitch was thrown}
\item{pitch_type}{Type of pitch thrown:  CH (Change-up), CU (Curveball), FF (Four-Seam Fastball),
FT (Two-Seam Fastball), SL (Slider).  Pitch type is determined by computer algorithm.}
\item{balls}{Current ball count}
\item{strikes}{Current strike count}
\item{pitches}{number of pitches previously thrown in the game}
\item{speed}{speed of pitch (in mph).  (When crossing plate?)}
\item{px}{x-coordinate of pitch (in feet, measured from center of plate)}
\item{pz}{vertical coordinate of pitch (in feet above plate)}
\item{pfx_x}{the horizontal movement, in inches, of the pitch between the release point and
home plate, as compared to a theoretical pitch thrown at the same speed with no spin-induced
movement. Measured at 40 feet from home plate.}
\item{pfx_z}{the vertical movement, in inches, of the pitch between the release point and
home plate, as compared to a theoretical pitch thrown at the same speed with no spin-induced
movement. Measured at 40 feet from home plate.}
\item{batter_hand}{A factor with two values: L (left) and R (right).}
 }}
\source{
Marchi and Albert:  Analyziing Baseball Data with R, CRC Press 2014.  For
more on the PITCHf/x system, see \url{http://en.wikipedia.org/wiki/PITCHf/x}.
}
\description{
PITCHf/x data on Justin Verlander, winner of the 2011 Cy Young Award.  Covers his
2009 through 2012 seasons.
}
\keyword{datasets}

