% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_statisticians.R
\name{watch_statisticians_slow}
\alias{watch_statisticians_slow}
\title{Multiple Confidence Interval for a Proportion, One at a Time}
\usage{
watch_statisticians_slow(n, p, interval_number = 50, level = 0.95)
}
\arguments{
\item{n}{the sample size}

\item{p}{the population proportion}

\item{interval_number}{number of intervals to make (limit and default are 50)}

\item{level}{desired level of confidence}
}
\value{
Graph side-effects
}
\description{
An app to explore confidence interval for a proportion.
}
\examples{
\dontrun{
watch_statisticians_slow(n = 100, p = 0.5, interval_number = 50, level = 0.95)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
