% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{youthrisk03}
\alias{youthrisk03}
\title{Youth Risk 2003}
\format{
A data frame with 3042 observations on 7 variables.
}
\source{
Mind on Statistics, 4th Edition, by Uts and Heckard.
}
\description{
A Study of Risky Behaviors in High School Seniors, from year 2003.

\itemize{
\item \code{Sex}
\item \code{Grades} Typical grades you earn in school
\item \code{WtAction} What do you plan to do about your weight?
\item \code{Seatbelt} How often do you wear a seat-belt?
\item \code{Sunscreen} How much do you wear sunscreen?
\item \code{Grades_1} Same as grades, but with some groups combined
\item \code{Sun_1} Same as sunscreen, but with some groups combined
}
}
\keyword{datasets}
