\docType{data}
\name{hanford1}
\alias{hanford1}
\title{Handford Weather Station, 1984-2010}
\format{A data frame with 27 observations on the following 2 variables.
\describe{
\item{year}{calendar year}
\item{temp}{average high temperature for that year.}
}}
\source{
See \url{http://www.hanford.gov/hms/}
}
\description{
The station is located in Hanford, WA.
}
\keyword{datasets}

