% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGC.R
\name{helpGC}
\alias{helpGC}
\title{Quick Vignettes in the Viewer}
\usage{
helpGC(topic,package="tigerstats")
}
\arguments{
\item{topic}{filename of the vignette, exclusive of the .html extension}

\item{package}{Name of the installed package containing the desried vignette}
}
\value{
side effects
}
\description{
A convenience function to show package vignettes.  Vignette will show in a 
viewer selected by the front-end (e.g, the Viewer pane in R Studio), or in a 
browser if the "viewer" option is not set.
}
\examples{
\dontrun{
helpGC(lmGC)
}
}
\author{
Homer White (hwhite0@georgetowncollege.edu)
}

