% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpGC.R
\name{helpGC}
\alias{helpGC}
\title{Extensive Help Via Vignettes}
\usage{
helpGC(topic)
}
\arguments{
\item{topic}{filename of the vignette, exclusive of the .html extension}
}
\value{
side effects
}
\description{
A convenience function to show vignettes associated with
package \code{tigerstats}.  Vignette will open in the user's default
browser.
}
\examples{
\dontrun{
helpGC(lmGC)
}
}
\author{
Homer White (hwhite0@georgetowncollege.edu)
}

