% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmGC.R
\name{lmGC}
\alias{lmGC}
\title{Linear Regression}
\usage{
lmGC(form,data=parent.frame(),graph=FALSE,check=FALSE)
}
\arguments{
\item{form}{formula of form y~x, both variables numeric}

\item{data}{dataframe supplying y and x above.  If one or more of the variables is not in data, then
they will be searched for in the parent environment.}

\item{graph}{Produce scatterplot with fitted ploynomial, together with prediction standard error bands}

\item{check}{Asks to produce a lowess or gam curve with approximate 95%-confidence band.  If the
fitted line wanders outside the band, then perhaps a linear fit is not appropriate.}
}
\value{
A list of class "GClm".  Elements that may be queried include "slope", "intercept",
"s" (residual standard error), "R^2" (unadjusted).
}
\description{
Regression analysis (one numerical predictor variable) with simplified output.
  Wrapper function for \code{lm} in package \code{stats}.
}
\examples{
#To study the relationship between two numerical variables:
lmGC(fastest~GPA,data=m111survey,graph=TRUE)
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

