% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m111survey}
\alias{m111survey}
\title{MAT 111 Survey}
\format{A data frame with 71 rows and 12 variables}
\source{
Georgetown College, MAT 111.
}
\description{
Results of a survey of MAT 111 students at Georgetown College.

\itemize{
      \item height.  How tall are you, in inches?
      \item ideal_ht.  A numeric vector  How tall would you LIKE to be, in inches?
      \item sleep.  How much sleep did you get last night?
      \item fastest.  What is the highest speed at which you have ever driven a car?
      \item weight_feel.  How do you feel about your weight?
      \item love_first.  Do you believe in love at first sight?
      \item extra_life.  Do you believe in extraterrestrial life?
      \item seat.  When you have a choice, where do you prefer to sit in a classroom?
      \item GPA.  What is your college GPA?
      \item enough_Sleep.  Do you think you get enough sleep?
      \item sex.  What sex are you?
      \item diff.  Your ideal height minus your actual height.
}
}
\keyword{datasets}

