% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{ParseFormula}
\alias{ParseFormula}
\title{Parse formulas}
\usage{
ParseFormula(formula,...)
}
\arguments{
\item{formula,}{a formula}

\item{\dots}{additional arguments, should folks decide to add them someday}
}
\value{
an object of class \code{parsedFormula}, used to compute on the language
}
\description{
utility for extracting portions of formulas.
}
\author{
Inspired by similar function in package \code{mosaic}.
Included in this package to reduce dependency.
}
\keyword{internal}

