\name{plot.GClm}
\alias{plot.GClm}
\title{Print Function for GC Linear Regression}
\usage{
  \S3method{plot}{GClm}(x,y,...)
}
\arguments{
  \item{x}{An object of class GCttest.}

  \item{y}{ignored}

  \item{\ldots}{ignored}

  \item{GClm}{an object of class GClm}
}
\value{
  graphical output
}
\description{
  Utility diagnostic plot
}
\author{
  Homer White \email{hwhite0@georgetowncollege.edu}
}

