% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pushups}
\alias{pushups}
\title{Pushups by Football Players at Georgetown College}
\format{A data frame with 30 observations on the following 3 variables.
\describe{
\item{weight}{weight of subject in
pounds.}
\item{pushups}{number of pushups
completed.}
\item{position}{a factor with levels \code{LINE}
\code{SKILL}: type of position played by the subject.  Line positions
require high body mass, skill positions require a lot of running around.}
}}
\source{
MAT 111, Georgetown College
}
\description{
Two football players at GC asked their team-mates to do as many pushups as
they could in two minutes.
}
\keyword{datasets}

