% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{RectShade}
\alias{RectShade}
\title{Shade Rectangles for Discrete Distributions}
\usage{
RectShade(low,high,func,...)
}
\arguments{
\item{low}{lower bound}

\item{high}{upper bound}

\item{func}{probability mass function}

\item{\ldots}{other arguments passed (to modify func)}
}
\value{
graphical side effect only
}
\description{
Utility function for pbinomGC ...
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
\keyword{internal}

