% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampDistMean.R
\name{SampDistMean}
\alias{SampDistMean}
\title{Distribution of the Sample Mean}
\usage{
SampDistMean(pop,max.samp.size=50,sim.reps=1000)
}
\arguments{
\item{pop}{A data frame representing the population from which samples are taken.}

\item{max.samp.size}{Largest sample size shown on the slider.}

\item{sim.reps}{Number of simulation repetitions to construct empirical distribution of the sample mean.}
}
\value{
Graphical and numerical output.
}
\description{
An app to explore the Central Limit Theorem.
}
\note{
Uses \code{manipulate} in R Studio.
}
\examples{
\dontrun{
data(imagpop)
if (require(manipulate)) SampDistMean(imagpop)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
