% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleKeyRev}
\alias{simpleKeyRev}
\title{Reversed Simple Key Function

Utility function for barchartGC.}
\usage{
simpleKeyRev(text, points = TRUE,
rectangles = FALSE,
lines = FALSE,
col = add.text$col,
cex = add.text$cex,
alpha = add.text$alpha,
font = add.text$font,
fontface = add.text$fontface,
fontfamily = add.text$fontfamily,
lineheight = add.text$lineheight,
...)
}
\description{
Reversed Simple Key Function

Utility function for barchartGC.
}
\keyword{internal}

