% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchartGC.R
\name{simpleKeyRev}
\alias{simpleKeyRev}
\title{Reversed Simple Key Function}
\usage{
simpleKeyRev(text, points = TRUE,
rectangles = FALSE,
lines = FALSE,
col = add.text$col,
cex = add.text$cex,
alpha = add.text$alpha,
font = add.text$font,
fontface = add.text$fontface,
fontfamily = add.text$fontfamily,
lineheight = add.text$lineheight,
...)
}
\description{
utility function for barchartGC
}
\keyword{internal}
