% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleRandom2.R
\name{SimpleRandom2}
\alias{SimpleRandom2}
\title{Simple Random Sample (v2)}
\usage{
SimpleRandom2(n=100)
}
\arguments{
\item{n}{the desired sample size}
}
\value{
Graphical and numerical output
}
\description{
Similar to \code{\link{SimpleRandom}}, but with a 
fixed sample size set by the user.
}
\examples{
\dontrun{
if (require(manipulate)) SimpleRandom()
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}
