\name{Skewer}
\alias{Skewer}
\title{SkewR}
\usage{
Skewer()
}
\value{
Graphical output.
}
\description{
An app to illustrate the effect of skewness on the shape of
a boxplot.
}
\note{
Requires \code{manipulate}; uses functions from package
\code{lattice}
}
\examples{
\dontrun{
if (require(mainpulate)) Skewer()
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

