\docType{data}
\name{SmallExp}
\alias{SmallExp}
\title{A Small Experiment}
\format{A data frame with 16 observations on the following 3 variables.
\describe{ \item{name}{name of the subject}
\item{sex}{sex of the subject}
\item{athlete}{whether or not the subject is an athlete} }}
\description{
Subjects in a hypothetical experiment
}
\keyword{datasets}

