% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme.rpres}
\alias{theme.rpres}
\title{Lattice Theme or R Presentations}
\usage{
theme.rpres()
}
\value{
Returns a list to be supplied as the \code{theme} to the \code{lattice} function
\code{\link[lattice]{trellis.par.set}()}.
}
\description{
Modifies the current theme for use with lattice graphics in R Presentation dicuments. Increases size of title,
 axis lables and axis numbers, thickens some lines, etc.
}
\note{
Deprecated in favor of \code{themerpres()}.  May not appear in future versions.
}
\examples{
trellis.par.set(theme=theme.rpres())

}
\seealso{
\code{\link[lattice]{trellis.par.set}}, \code{\link[lattice]{show.settings}}
}
\keyword{graphics}
