% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigerstats-package.R
\docType{package}
\name{tigerstats-package}
\alias{tigerstats-package}
\alias{tigerstats}
\title{tigerstats}
\description{
Datasets and functions useful for teaching elementary statistics.
}
\references{
\url{https://homerhanumat.github.io/tigerstats/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/homerhanumat/tigerstats}
  \item Report bugs at \url{https://github.com/homerhanumat/tigerstats/issues}
}

}
\author{
Rebekah Robinson (\email{rebekahmusselwhite23@gmail.com}), Homer White (\email{hwhite0@georgetowncollege.edu})
}
\keyword{package}
