\docType{package}
\name{tigerstats-package}
\alias{tigerstats}
\alias{tigerstats-package}
\title{R for Elementary Statistics}
\description{
tigerstats
}
\details{
Datasets and functions useful for teaching elementary
statistics.
}
\author{
Rebekah Robinson
(\email{Rebekah_Robinson@georgetowncollege.edu}), Homer
White (\email{hwhite0@georgetowncollege.edu})
}
\references{
\url{http://statistics.georgetowncollege.edu}
}
\keyword{package}

