% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigerstats-package.R
\docType{package}
\name{tigerstats-package}
\alias{tigerstats-package}
\alias{tigerstats}
\title{R Functions for Elementary Statistics}
\description{
tigerstats
}
\details{
Datasets and functions useful for teaching elementary statistics.
}
\references{
\url{http://statistics.georgetowncollege.edu}
}
\author{
Rebekah Robinson (\email{rebekahrobinson14@gmail.com}), Homer White (\email{hwhite0@georgetowncollege.edu})
}
\keyword{package}
