\name{tigerstats-package}
\alias{tigerstats-package}
\alias{tigerstats}
\docType{package}
\title{
R for Statistics at Georgetown College
}
\description{
Datasets and functions useful for teaching elementary statistics.
}
\details{
\tabular{ll}{
Package: \tab tigerstats\cr
Type: \tab Package\cr
Version: \tab 0.1.2-7\cr
Date: \tab 2014-03-08\cr
License: \tab GPL (>=3)\cr
}

}
\author{
Homer White and Rebekah Robinson

Maintainer: Homer White \email{hwhite0@georgetowncollege.edu}
}

\keyword{ package }
