\name{tigerstats-package}
\alias{tigerstats-package}
\alias{tigerstats}
\docType{package}
\title{
R for Elementary Statistics
}
\description{
Datasets and functions useful for teaching elementary statistics.
}
\details{
\tabular{ll}{
Package: \tab tigerstats\cr
Type: \tab Package\cr
Version: \tab 0.1.4-1\cr
Date: \tab 2014-04-11\cr
License: \tab GPL (>=3)\cr
}

}
\author{
Homer White and Rebekah Robinson

Maintainer: Homer White \email{hwhite0@georgetowncollege.edu}
}

\keyword{ package }
